+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.16.2
+
+* added optional argument to `File.get_value`